collectors/exchange-bybit-history-data/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.0
pytest==8.0.0